check other profile add form modification hike

+++ src/views/profile.view.php
@@ -1,22 +1,34 @@
+<?php extract($user); ?>
 <section>
-    <h2>Ton profil</h2>
+    <?php if ($_SESSION['hiking_user']['uid'] == $uid): ?>
+        <h2>Ton profile</h2>
+    <?php else: ?>
+        <h2>Profile de <?= $nickname ?></h2>
+    <?php endif; ?>
     <ul>
-        <li>Nom: <?= $_SESSION['hiking_user']['firstname'] . " " . $_SESSION['hiking_user']['lastname'] ?></li>
-        <li>Pseudo: <?= $_SESSION['hiking_user']['nickname'] ?></li>
-        <li>Email: <?= $_SESSION['hiking_user']['email'] ?></li>
+        <li>Nom: <?= $firstname . " " . $lastname ?></li>
+        <li>Pseudo: <?= $nickname ?></li>
+        <li>Email: <?= $email ?></li>
     </ul>
 </section>
 <section>
-    <h2>Tes Hikes</h2>
+    <?php if ($_SESSION['hiking_user']['uid'] == $uid): ?>
+        <h2>Tes Hikes</h2>
+    <?php else: ?>
+        <h2>Hikes de <?= $nickname ?></h2>
+    <?php endif; ?>
     <ul>
         <?php if (!empty($hikes)):
             foreach ($hikes as $hike):
                 extract($hike); ?>
                 <li>
-                    <a href="/hike?hid=<?= $hid ?>"><?= $name ?></a>
-                    <a href="/modify?value=hike"><i class="fa-solid fa-pencil"></i></a>
-                    <a href="/delete-hike?hid=<?= $hid ?>" onclick="return confirm('Etes-vous sûr de supprimer cette randonnée ?')"><i class="fa-regular fa-trash-can"></i></a>
-
+                    <?php if ($_SESSION['hiking_user']['uid'] == $uid): ?>
+                        <a href="/hike?hid=<?= $hid ?>"><?= $name ?></a>
+                        <a href="/modify?value=hike&hid=<?= $hid ?>"><i class="fa-solid fa-pencil"></i></a>
+                        <a href="/delete-hike?hid=<?= $hid ?>" onclick="return confirm('Etes-vous sûr de supprimer cette randonnée ?')"><i class="fa-regular fa-trash-can"></i></a>
+                    <?php else: ?>
+                        <a href="/hike?hid=<?= $hid ?>"><?= $name ?></a>
+                    <?php endif; ?>
                 </li>
             <?php endforeach;
             else: ?>
